Fixing out-of-range bugs found during testing
- Fixed bugs that were happening for some edge cases:
1) Records without any record separators were causing findFirstMatch to
hit out-of-bounds errors (since `endOfKey` would be -1 in that case).
Fixed by checking for that and falling back to either the end of the
current record or the end of the DB.
2) RangeSearch called with `endNeedle < startNeedle` caused a panic, as
it should. But it’s preferable for the server to handle this issue more
gracefully and just return an empty set. Added a test for this case as
well.
3) Fixed mismatches at the start of ranges. These were sometimes
happening as a result of the assumption that a record separator is
appended to both `startNeedle` and `endNeedle`. While this is a valid
assumption for exact matching (i.e. db.Search), it messes up range
comparisons and causes some misses that shouldn’t be happening. Updated
documentation of range match to remove this requirement.